Move scraper to lib/ to avoid Zeitwerk autoloading it
It was causing issues in prod because capybara is not installed in prod
